requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
yfinance>=0.2.0
nsepy>=0.8
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# NLP and sentiment analysis
nltk>=3.8
textblob>=0.17.0
transformers>=4.30.0
torch>=2.0.0

# Technical indicators
ta-lib>=0.4.0
ta>=0.11.0

# Database and storage
sqlalchemy>=2.0.0
pymongo>=4.5.0

# API and web scraping
newspaper3k>=0.2.8
feedparser>=6.0.10

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
